requirements.txt
Youtube_transcript_api
streamlit
google-generativeai
python-dotenv
pathlib